+++ changelog.md
@@ -7,6 +7,7 @@
   + `no-implicit-any`: renamed to `no-unknown`
 * `CHG` formatter: no longer need` --preview`
 * `FIX` semantic: color of `function`
+* `FIX` [#1027](https://github.com/sumneko/lua-language-server/issues/1027)
 * `FIX` [#1028](https://github.com/sumneko/lua-language-server/issues/1028)
 
 ## 2.6.8