+++ changelog.md
@@ -1,5 +1,8 @@
 # changelog
 
+## 3.0.1
+* `FIX` [#1033](https://github.com/sumneko/lua-language-server/issues/1033)
+
 ## 3.0.0
 `2022-4-10`
 * `CHG` [break changes](https://github.com/sumneko/lua-language-server/issues/980)